Created a linked list and node data structure class
To try to optimize I would like to try to store the vertices that we calculate in mountain.
My goal is to only calculate enough noise and map to fill the screen in setup then each
frame after only one noise and map need to be done since the old values are already
computed. This should speed up the draw function in mountain.js