Implemented a initialize function that creates the vertices and a new draw function
This update to the code does not seem to make the sketch run more smoothly. I believe now
the problem is that are are over 20000 vertices and that is slowing the main draw function.
A possible solution would be to draw every other or even every third vertex. This
interpolation would make the curves of the mountain less smooth, but since there are so
many vertices it would hopefully be not very noticable